pyproject.toml
[project]
name = "chatbot-external-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.44.1",
    "langchain-core>=0.3.75",
    "langchain-google-genai>=2.1.10",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "langgraph-checkpoint-postgres>=2.0.23",
    "psycopg[binary,pool]>=3.2.9",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
